+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All Notable changes to the `digipolisgent/values` package.
 
-## [Unreleased]
+## [1.0.0]
 
-[Unreleased]: https://github.com/digipolisgent/php_package_dg-values/compare/master...develop
+### Added
+
+- Added the ValueInterface.
+- Added the ValueFromArrayInterface.
+- Added the ValueAbstract class.
+- Added the CollectionInterface.
+- Added the CollectionAbstract class.
+
+[1.0.0]: https://github.com/digipolisgent/php_package_dg-value/tree/1.0.0
+[Unreleased]: https://github.com/digipolisgent/php_package_dg-value/compare/master...develop